karla/pyproject.toml
[project]
name = "karla"
version = "0.1.0"
description = "A Python coding agent with Letta backend and ACP support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "letta-client",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
acp = [
    "agent-client-protocol>=0.1.0",
]
evals = [
    "letta-evals",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.uv.sources]
letta-client = { git = "https://github.com/odellus/letta-python.git", rev = "5a1c107553619dacac5747b08af504eef9487647" }
letta-evals = { git = "https://github.com/odellus/letta-evals.git", rev = "f22f8fd8500041bcb4407a231590868e7af46cf6" }

[project.scripts]
karla = "karla.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/karla"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (require running Letta server)",
]
